pyproject.toml
[project]
name = "storefront"
version = "0.1.0"
description = "A toolkit for embedding interactive storefront design to your product"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.4",
    "python-liquid>=2.1.0",
    "werkzeug>=3.1.3",
]
